Implement new classes and functions for several modules
In this commit, Config, IntResult, BankAccount, and IntMatrix classes were implemented as per requirements. For the Config part, it was done to set up to parse Config from String and provide safe access to its keys. For IntResult class, is was implemented to handle integer results with cases of value or error. BankAccount got functionality like: deposit, withdraw, and balance display. IntMatrix is now a valid two-dimensional integer array-like class with corresponding manipulation methods.