Refactor IntMatrix backing structure to flat array
Changed the backingMatrix in IntMatrix.kt from a 2D array to a single 1D array. This modification was done to improve space efficiency and read/write operations. Instead of using arrays of arrays we are now using the concept of a 'flat' array, saving space in memory and improving the speed of access and alteration.